Save Button Code Mistake Fix

+++ 2 - Javascript Practice Projects/17 - JavaScript FlashCard OOP Project Prototype/script.js
@@ -751,26 +751,28 @@ $(function () {
     function saveButton() {
         $("#saveButton").on({
             mouseup: function () {
-                questionValue = $("#questionTextArea").val();
-                answerValue = $("#answerTextArea").val();
-                inputValue = $("#type").val();
-                myObj = [inputValue, questionValue, answerValue];
-                stringfy = JSON.stringify(myObj);
-                $(this).css("background-color", "var(--emptyColor)");
-                tempObj.push(stringfy);
-                tempObj = $.unique(tempObj.sort());
-                let x = 0;
-                while (x < tempObj.length) {
-                    localStorage.setItem(x, tempObj[x])
-                    x++;
+                if ($("#questionTextArea").val().trim() != "" && $("#answerTextArea").val().trim() != "") {
+                    questionValue = $("#questionTextArea").val();
+                    answerValue = $("#answerTextArea").val();
+                    inputValue = $("#type").val();
+                    myObj = [inputValue, questionValue, answerValue];
+                    stringfy = JSON.stringify(myObj);
+                    $(this).css("background-color", "var(--emptyColor)");
+                    tempObj.push(stringfy);
+                    tempObj = $.unique(tempObj.sort());
+                    let x = 0;
+                    while (x < tempObj.length) {
+                        localStorage.setItem(x, tempObj[x])
+                        x++;
+                    }
+                    x = 0
+                    console.log(localStorage)
+                    $(".card0Div").empty();
+                    questionAnswerInner();
+                    i = 0;
+                    showHide();
+                    deleteClass();
                 }
-                x = 0
-                console.log(localStorage)
-                $(".card0Div").empty();
-                questionAnswerInner();
-                i = 0;
-                showHide();
-                deleteClass();
             }
         })
     }
@@ -1167,11 +1169,7 @@ $(function () {
         $(".hideDeleteContainer").css("grid-template-areas", "hideButton");
         $("#delete").css("display", "none");
         $("#deleteAll").css("display", "none");
-        $("#saveButton").on({
-            mouseup: function () {  
-                $(".card0Div").css("display", "none");
-            }
-        });
+        $(".card0Div").css("display", "none");
         $("#questionTextArea, #answerTextArea, #type").val("");
     }
     $("#addFlashCard").on({
